Process for harvesting CRC collections into ScienceBase
This commit includes a notebook and related functions that harvest items from the USGS Core Research Center into ScienceBase via ArcGIS service and web scraping. It includes significant notes about this process as an example for reference. Code is running live at the point of this commit, so all the blocks in the notebook do not show latest output.